Complete fix bug datepicker not working when page first load by removing turbolinks gem, then refactor code to StimulusJS date_picker controller
