+++ app/controllers/reservation_controller.rb
@@ -1,4 +1,13 @@
 class PropertyController < ApplicationController
+  before_action :authenticate_user!
+
   def create 
+    Reservation.create(reservation_params)
+  end
+
+  private 
+
+  def reservation_params 
+    
   end
 end
