+++ app/javascript/controllers/datepicker_controller.js
@@ -1,0 +1,12 @@
+import { Controller } from "@hotwired/stimulus"
+import flatpickr from "flatpickr"
+import 'flatpickr/dist/flatpickr.min.css'
+
+export default class extends Controller {
+  connect() {
+    flatpickr(".datepicker", {
+      minDate: "today",
+      showMonths: 2
+    });
+  }
+}
