+++ app/views/owner/listings/_header.slim
@@ -1,11 +1,10 @@
-nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36[data-controller="home"]
+nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36 [data-controller="dropdown"]
   .flex.items-center.justify-between.relative
     = link_to root_path do
       .text-pinkRed
         = svg_pack_tag 'airbnb_brand', alt: "Airbnb Brand"
 
-    = form_with url: owner_start_build_property_path, method: :post do |f|
-        = f.submit "Create new listing", class: "text-gray-500 px-5 py-3 rounded-lg font-semibold hover:text-black"
+    = link_to "Create new listing", owner_start_build_property_path, class: "text-gray-500 px-5 py-3 rounded-lg font-semibold hover:text-black"
     
     .flex.gap-5.items-center        
       .relative.header__dropdown [data-dropdown]
@@ -26,35 +25,3 @@ nav.sticky.top-0.left-0.w-screen.pt-4.bg-white.lg:px-36[data-controller="home"]
             = link_to "Switch to traveling", root_path
             = link_to "Logout", destroy_user_session_path, data: { turbo_method: :delete } 
 
-
-  .hidden.fixed.inset-0.bg-black.bg-opacity-60.transition-opacity [data-home-target="modal"]
-    .flex.min-h-full.items-end.justify-center.p-4.text-center.sm:items-center.sm:p-0
-      .relative.transform.overflow-hidden.rounded-lg.bg-white.text-left.pb-4.shadow-xl.transition-all.sm:my-8.sm:w-full.sm:max-w-lg
-        .relative.px-4.pt-5.pb-4.sm:p-6.sm:pb-4
-          button.absolute.top-4.left-4.rounded-full.p-2[data-action="click->home#closeModal"]
-            = svg_pack_tag "close_icon"
-          h3#modal-title.text-lg.leading-6.text-center
-            | Log in or sign up
-          .
-        
-        h1.px-4.text-xl.mb-4.mt-7
-          |Welcome to Airbnb
-        
-        form.px-4.flex.flex-col.gap-4.text-black [data-controller="authentication"]
-          input.py-3.border.border-gray-400.rounded-lg type="email" placeholder="Email" data-authentication-target="email" class="focus:border-2 focus:border-black focus:ring-0" 
-
-          button.inline-flex.w-full.justify-center.rounded-md.border.border-transparent.bg-pinkRed.px-4.py-3.text-base.font-bold.text-white.shadow-sm [data-action="click->authentication#submitForm"]
-            | Continue
-              
-        .px-4.py-3.flex.flex-col.gap-4
-          p.text-center.text-gray-500
-            |or
-          button.inline-flex.w-full.justify-between.items-center.rounded-md.border.border-gray-300.bg-white.px-4.py-3.text-base.text-black.shadow-sm.border.border-black
-            = svg_pack_tag "facebook_icon"
-            | Continue with Facebook
-            .
-          
-          button.inline-flex.w-full.justify-between.items-center.rounded-md.border.border-gray-300.bg-white.px-4.py-3.text-base.text-black.shadow-sm.border.border-black
-            = svg_pack_tag "google_icon"
-            | Continue with Google
-            .
