+++ app/views/property/show.slim
@@ -1,4 +1,4 @@
-.w-screen.px-96
+.w-screen.px-96 [data-controller="datepicker"]
   = render partial: "header"
 
   .mt-16
@@ -21,7 +21,7 @@
         p.text-lg.text-gray-700
           = @property.description
 
-      .sticky.border.border-gray-300.rounded-xl.right-0.top-28.p-8.flex.flex-col.gap-4 class="w-5/12"
+      .sticky.border.border-gray-300.rounded-xl.right-0.top-28.p-8.flex.flex-col.gap-4 class="w-5/12" 
         h3.text-lg
           span.font-bold.text-2xl 
             |$999
