+++ package.json
@@ -8,8 +8,7 @@
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
     "@rails/webpacker": "4.3.0",
-    "flatpickr": "^4.6.13",
-    "turbolinks": "^5.2.0"
+    "flatpickr": "^4.6.13"
   },
   "version": "0.1.0",
   "devDependencies": {
